requirements.txt
streamlit==1.28.0
plotly==5.17.0
pandas==2.0.3
sqlite3
gspread==5.12.0
oauth2client==4.1.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
qrcode==7.4.2
Pillow==10.1.0
streamlit-option-menu==0.3.6
pyzbar==0.1.9
opencv-python==4.8.1.78